# analytics/contracts/http/openapi.yaml
openapi: 3.1.0
info:
  title: Analytics HTTP API
  version: "1.0.0"
  description: >
    Canonical HTTP contract. Request bodies for ingest $ref the JSON Schemas
    in schemas/events/. Read-API responses $ref schemas/responses/.
    Shopware backend events are not accepted on HTTP ingest; they are
    documented in contracts/rabbitmq/asyncapi.yaml.
servers:
  - url: /api/v1
paths:
  /events:
    post:
      operationId: ingestEvent
      tags: [Ingestion]
      summary: Ingest a single frontend event
      security:
        - ingestBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FrontendEvent"
            examples:
              session_started:
                $ref: "./examples/valid/session-started.json"
              product_viewed:
                $ref: "./examples/valid/product-viewed.json"
              add_to_cart:
                $ref: "./examples/valid/add-to-cart.json"
              checkout_started:
                $ref: "./examples/valid/checkout-started.json"
              payment_methods_shown:
                $ref: "./examples/valid/payment-methods-shown.json"
              payment_method_selected:
                $ref: "./examples/valid/payment-method-selected.json"
              payment_failed:
                $ref: "./examples/valid/payment-failed.json"
              contact_intent:
                $ref: "./examples/valid/contact-intent.json"
      responses:
        "200":
          description: Accepted or duplicate of the same event.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestEventResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Same event_id already stored with a different event.
        "422":
          description: Schema, allowlist, or unknown sales_channel_id.
        "429":
          description: HTTP ingest rate limit exceeded. Retry-After is in seconds.
  /events/batch:
    post:
      operationId: ingestEventBatch
      tags: [Ingestion]
      summary: Ingest a batch of frontend events
      security:
        - ingestBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FrontendEventBatch"
      responses:
        "200":
          description: Per-item accepted / duplicate / rejected. One item failure does not hide the rest.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestBatchResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Same event_id already stored with a different event.
        "422":
          description: Batch envelope invalid (item-level 422 is returned inside 200).
        "429":
          description: HTTP ingest rate limit exceeded. Retry-After is in seconds.
  /analytics/overview:
    get:
      operationId: getOverview
      tags: [Reports]
      summary: Overview metrics for a period
      security:
        - readBearer: []
      parameters:
        - $ref: "#/components/parameters/PeriodFrom"
        - $ref: "#/components/parameters/PeriodTo"
        - $ref: "#/components/parameters/SalesChannel"
        - $ref: "#/components/parameters/Market"
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/Campaign"
        - $ref: "#/components/parameters/Sku"
        - $ref: "#/components/parameters/Channel"
        - $ref: "#/components/parameters/PaymentMethod"
        - $ref: "#/components/parameters/Currency"
        - $ref: "#/components/parameters/AttributionModel"
      responses:
        "200":
          description: Overview counts, money by currency, and conversion rates.
          content:
            application/json:
              schema:
                $ref: "./schemas/responses/overview.schema.json"
              examples:
                googleAdsDe:
                  $ref: "./examples/responses/overview.json"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /analytics/overview/daily:
    get:
      operationId: getOverviewDaily
      tags: [Reports]
      summary: Daily paid sales and revenue
      security:
        - readBearer: []
      parameters:
        - $ref: "#/components/parameters/PeriodFrom"
        - $ref: "#/components/parameters/PeriodTo"
        - $ref: "#/components/parameters/SalesChannel"
        - $ref: "#/components/parameters/Market"
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/Campaign"
        - $ref: "#/components/parameters/Sku"
        - $ref: "#/components/parameters/Channel"
        - $ref: "#/components/parameters/PaymentMethod"
        - $ref: "#/components/parameters/Currency"
        - $ref: "#/components/parameters/AttributionModel"
      responses:
        "200":
          description: Paid sales and money by UTC day.
          content:
            application/json:
              schema:
                $ref: "./schemas/responses/overview-daily.schema.json"
              examples:
                googleAdsDe:
                  $ref: "./examples/responses/overview-daily.json"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /analytics/funnel:
    get:
      operationId: getFunnel
      tags: [Reports]
      summary: Ecommerce and Lead funnels
      security:
        - readBearer: []
      parameters:
        - $ref: "#/components/parameters/PeriodFrom"
        - $ref: "#/components/parameters/PeriodTo"
        - $ref: "#/components/parameters/SalesChannel"
        - $ref: "#/components/parameters/Market"
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/Campaign"
        - $ref: "#/components/parameters/Sku"
        - $ref: "#/components/parameters/Channel"
        - $ref: "#/components/parameters/PaymentMethod"
        - $ref: "#/components/parameters/Currency"
        - $ref: "#/components/parameters/AttributionModel"
      responses:
        "200":
          description: Entity counts per step and adjacent conversion.
          content:
            application/json:
              schema:
                $ref: "./schemas/responses/funnel.schema.json"
              examples:
                googleAdsDe:
                  $ref: "./examples/responses/funnel.json"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /analytics/sources:
    get:
      operationId: getSources
      tags: [Reports]
      summary: Sources and campaigns
      security:
        - readBearer: []
      parameters:
        - $ref: "#/components/parameters/PeriodFrom"
        - $ref: "#/components/parameters/PeriodTo"
        - $ref: "#/components/parameters/SalesChannel"
        - $ref: "#/components/parameters/Market"
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/Campaign"
        - $ref: "#/components/parameters/Sku"
        - $ref: "#/components/parameters/Channel"
        - $ref: "#/components/parameters/PaymentMethod"
        - $ref: "#/components/parameters/Currency"
        - $ref: "#/components/parameters/AttributionModel"
      responses:
        "200":
          description: Rows keyed by source and campaign for the selected attribution model.
          content:
            application/json:
              schema:
                $ref: "./schemas/responses/sources.schema.json"
              examples:
                googleAdsDe:
                  $ref: "./examples/responses/sources.json"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /analytics/contact-channels:
    get:
      operationId: getContactChannels
      tags: [Reports]
      summary: Form, email, WhatsApp and phone channels
      security:
        - readBearer: []
      parameters:
        - $ref: "#/components/parameters/PeriodFrom"
        - $ref: "#/components/parameters/PeriodTo"
        - $ref: "#/components/parameters/SalesChannel"
        - $ref: "#/components/parameters/Market"
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/Campaign"
        - $ref: "#/components/parameters/Sku"
        - $ref: "#/components/parameters/Channel"
        - $ref: "#/components/parameters/PaymentMethod"
        - $ref: "#/components/parameters/Currency"
        - $ref: "#/components/parameters/AttributionModel"
      responses:
        "200":
          description: Contacts and unique Leads are counted separately.
          content:
            application/json:
              schema:
                $ref: "./schemas/responses/contact-channels.schema.json"
              examples:
                formAndWhatsapp:
                  $ref: "./examples/responses/contact-channels.json"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /analytics/products:
    get:
      operationId: getProducts
      tags: [Reports]
      summary: SKU product report
      security:
        - readBearer: []
      parameters:
        - $ref: "#/components/parameters/PeriodFrom"
        - $ref: "#/components/parameters/PeriodTo"
        - $ref: "#/components/parameters/SalesChannel"
        - $ref: "#/components/parameters/Market"
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/Campaign"
        - $ref: "#/components/parameters/Sku"
        - $ref: "#/components/parameters/Channel"
        - $ref: "#/components/parameters/PaymentMethod"
        - $ref: "#/components/parameters/Currency"
        - $ref: "#/components/parameters/AttributionModel"
      responses:
        "200":
          description: Views, carts, paid orders, revenue and payment methods per SKU.
          content:
            application/json:
              schema:
                $ref: "./schemas/responses/products.schema.json"
              examples:
                sofa:
                  $ref: "./examples/responses/products.json"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /analytics/payment-methods:
    get:
      operationId: getPaymentMethods
      tags: [Reports]
      summary: Payment method availability, selection and conversion
      security:
        - readBearer: []
      parameters:
        - $ref: "#/components/parameters/PeriodFrom"
        - $ref: "#/components/parameters/PeriodTo"
        - $ref: "#/components/parameters/SalesChannel"
        - $ref: "#/components/parameters/Market"
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/Campaign"
        - $ref: "#/components/parameters/Sku"
        - $ref: "#/components/parameters/Channel"
        - $ref: "#/components/parameters/PaymentMethod"
        - $ref: "#/components/parameters/Currency"
        - $ref: "#/components/parameters/AttributionModel"
      responses:
        "200":
          description: selected_rate denominator is checkouts where the method was shown.
          content:
            application/json:
              schema:
                $ref: "./schemas/responses/payment-methods.schema.json"
              examples:
                paypal:
                  $ref: "./examples/responses/payment-methods.json"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /analytics/period-comparison:
    get:
      operationId: getPeriodComparison
      tags: [Reports]
      summary: Compare two periods
      description: >
        Returns absolute and percent change for overview metrics and payment
        methods. The comparison shows correlation, not causation.
      security:
        - readBearer: []
      parameters:
        - $ref: "#/components/parameters/PeriodFrom"
        - $ref: "#/components/parameters/PeriodTo"
        - $ref: "#/components/parameters/CompareFrom"
        - $ref: "#/components/parameters/CompareTo"
        - $ref: "#/components/parameters/SalesChannel"
        - $ref: "#/components/parameters/Market"
        - $ref: "#/components/parameters/Source"
        - $ref: "#/components/parameters/Campaign"
        - $ref: "#/components/parameters/Sku"
        - $ref: "#/components/parameters/Channel"
        - $ref: "#/components/parameters/PaymentMethod"
        - $ref: "#/components/parameters/Currency"
        - $ref: "#/components/parameters/AttributionModel"
      responses:
        "200":
          description: Current period, previous period, and deltas.
          content:
            application/json:
              schema:
                $ref: "./schemas/responses/period-comparison.schema.json"
              examples:
                visitorsUp:
                  $ref: "./examples/responses/period-comparison.json"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /analytics/journey/search:
    get:
      operationId: searchJourneys
      tags: [Journeys]
      summary: Search journeys by identifier or tracking value
      security:
        - readBearer: []
      parameters:
        - name: q
          in: query
          required: true
          description: >
            Visitor/session UUID, Shopware hex ID, order number, GCLID/GBRAID/WBRAID,
            campaign, contact channel, or tracking reference.
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Deduplicated entity hits, oldest matching events first.
          content:
            application/json:
              schema:
                $ref: "./schemas/responses/journey-search.schema.json"
              examples:
                visitorAndOrder:
                  $ref: "./examples/responses/journey-search.json"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /analytics/visitors/{visitor_id}/journey:
    get:
      operationId: getVisitorJourney
      tags: [Journeys]
      summary: Chronological visitor journey
      security:
        - readBearer: []
      parameters:
        - $ref: "#/components/parameters/VisitorId"
      responses:
        "200":
          $ref: "#/components/responses/Journey"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /analytics/visitors/{visitor_id}/anonymize:
    post:
      operationId: anonymizeVisitor
      tags: [Visitors]
      summary: Redact identifying visitor attributes
      security:
        - readBearer: []
      parameters:
        - $ref: "#/components/parameters/VisitorId"
      responses:
        "200":
          description: Visitor identifiers redacted. Rows are kept.
          content:
            application/json:
              schema:
                $ref: "./schemas/responses/anonymize-visitor.schema.json"
              examples:
                anonymized:
                  $ref: "./examples/responses/anonymize-visitor.json"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /analytics/leads/{lead_id}/journey:
    get:
      operationId: getLeadJourney
      tags: [Journeys]
      summary: Chronological lead journey
      security:
        - readBearer: []
      parameters:
        - $ref: "#/components/parameters/LeadId"
      responses:
        "200":
          $ref: "#/components/responses/Journey"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /analytics/orders/{order_id}/journey:
    get:
      operationId: getOrderJourney
      tags: [Journeys]
      summary: Chronological order journey
      security:
        - readBearer: []
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          $ref: "#/components/responses/Journey"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /analytics/customers/{customer_id}/journey:
    get:
      operationId: getCustomerJourney
      tags: [Journeys]
      summary: Chronological customer journey after customer_linked
      security:
        - readBearer: []
      parameters:
        - $ref: "#/components/parameters/CustomerId"
      responses:
        "200":
          $ref: "#/components/responses/Journey"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    ingestBearer:
      type: http
      scheme: bearer
      description: >
        Server-side ANALYTICS_INGEST_API_KEY for Next.js BFF / backend
        workers. Must not be embedded in browser JavaScript.
    readBearer:
      type: http
      scheme: bearer
      description: analytics_read_api_key
  parameters:
    PeriodFrom:
      name: period_from
      in: query
      required: true
      schema:
        type: string
        format: date-time
    PeriodTo:
      name: period_to
      in: query
      required: true
      schema:
        type: string
        format: date-time
    CompareFrom:
      name: compare_from
      in: query
      required: true
      schema:
        type: string
        format: date-time
    CompareTo:
      name: compare_to
      in: query
      required: true
      schema:
        type: string
        format: date-time
    SalesChannel:
      name: sales_channel
      in: query
      required: false
      schema:
        $ref: "./schemas/common/shopware-id.schema.json"
    Market:
      name: market
      in: query
      required: false
      schema:
        $ref: "./schemas/common/market-code.schema.json"
    Source:
      name: source
      in: query
      required: false
      schema:
        type: string
        minLength: 1
    Campaign:
      name: campaign
      in: query
      required: false
      schema:
        type: string
        minLength: 1
    Sku:
      name: sku
      in: query
      required: false
      schema:
        type: string
        minLength: 1
    Channel:
      name: channel
      in: query
      required: false
      schema:
        $ref: "./enums/contact-channel.schema.json"
    PaymentMethod:
      name: payment_method
      in: query
      required: false
      schema:
        type: string
        minLength: 1
    Currency:
      name: currency
      in: query
      required: false
      schema:
        type: string
        minLength: 3
        maxLength: 3
    AttributionModel:
      name: attribution_model
      in: query
      required: false
      schema:
        $ref: "./enums/attribution-model.schema.json"
    VisitorId:
      name: visitor_id
      in: path
      required: true
      schema:
        $ref: "./schemas/common/uuid-v4.schema.json"
    LeadId:
      name: lead_id
      in: path
      required: true
      schema:
        type: string
        minLength: 32
        maxLength: 32
        pattern: "^[0-9a-fA-F]{32}$"
    OrderId:
      name: order_id
      in: path
      required: true
      schema:
        type: string
        minLength: 32
        maxLength: 32
        pattern: "^[0-9a-fA-F]{32}$"
    CustomerId:
      name: customer_id
      in: path
      required: true
      schema:
        type: string
        minLength: 32
        maxLength: 32
        pattern: "^[0-9a-fA-F]{32}$"
  responses:
    Unauthorized:
      description: Missing or invalid API key.
    BadRequest:
      description: Invalid query, for example period_from > period_to.
    NotFound:
      description: Journey not found.
    Journey:
      description: Events ordered by occurred_at, then event_id.
      content:
        application/json:
          schema:
            $ref: "./schemas/responses/journey.schema.json"
          examples:
            visitorSession:
              $ref: "./examples/responses/journey.json"
  schemas:
    FrontendEvent:
      oneOf:
        - $ref: "./schemas/events/session-started.event.schema.json"
        - $ref: "./schemas/events/product-viewed.event.schema.json"
        - $ref: "./schemas/events/add-to-cart.event.schema.json"
        - $ref: "./schemas/events/checkout-started.event.schema.json"
        - $ref: "./schemas/events/payment-methods-shown.event.schema.json"
        - $ref: "./schemas/events/payment-method-selected.event.schema.json"
        - $ref: "./schemas/events/payment-failed.event.schema.json"
        - $ref: "./schemas/events/contact-intent.event.schema.json"
    FrontendEventBatch:
      type: object
      additionalProperties: false
      required: [events]
      properties:
        events:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/FrontendEvent"
    IngestEventResponse:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, duplicate]
    IngestBatchResponse:
      type: object
      additionalProperties: false
      required: [results]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/IngestBatchItem"
    IngestBatchItem:
      type: object
      additionalProperties: false
      required: [index, status]
      properties:
        index:
          type: integer
          minimum: 0
        event_id:
          oneOf:
            - type: string
            - type: "null"
        status:
          type: string
          enum: [accepted, duplicate, rejected]
        detail:
          type: string